Some XML documentation stuff.
Moved service interfaces and service client classes to OpenStory.ServiceModel, so they can be used without referencing the service assemblies. Which would hurt. Lots.
Renamed events to match the .NET convention, that is, without "On" at the start.
Renamed some event-related methods to also match the .NET convention, that is, with "On" at the start.